+++ pkg/band/downgrades.go
@@ -14,6 +14,8 @@
 
 package band
 
+import "go.thethings.network/lorawan-stack/pkg/ttnpb"
+
 // LoRaWAN 1.0.3rA -> 1.0.2rB downgrades
 
 func disableCFList1_0_2(b Band) Band {
@@ -43,7 +45,7 @@ func makeSetMaxTxPowerIndexFunc(idx uint8) func(Band) Band {
 // LoRaWAN 1.0.2rB -> 1.0.2rA downgrades
 
 func auDataRates1_0_2(b Band) Band {
-	for i := 0; i < 4; i++ {
+	for i := ttnpb.DataRateIndex(0); i < 4; i++ {
 		b.DataRates[i] = b.DataRates[i+2]
 	}
 	b.DataRates[5] = DataRate{}