+++ pkg/networkserver/grpc_asns.go
@@ -24,6 +24,7 @@ import (
 	pbtypes "github.com/gogo/protobuf/types"
 	"go.thethings.network/lorawan-stack/pkg/auth/rights"
 	"go.thethings.network/lorawan-stack/pkg/events"
+	"go.thethings.network/lorawan-stack/pkg/frequencyplans"
 	"go.thethings.network/lorawan-stack/pkg/log"
 	"go.thethings.network/lorawan-stack/pkg/rpcmetadata"
 	"go.thethings.network/lorawan-stack/pkg/ttnpb"
@@ -105,7 +106,7 @@ func (ns *NetworkServer) LinkApplication(link ttnpb.AsNs_LinkApplicationServer) 
 	return err
 }
 
-func validateApplicationDownlinks(session ttnpb.Session, macState *ttnpb.MACState, multicast bool, queue []*ttnpb.ApplicationDownlink, downs ...*ttnpb.ApplicationDownlink) (unmatchedQueue, unmatchedDowns []*ttnpb.ApplicationDownlink, err error) {
+func validateApplicationDownlinks(session ttnpb.Session, macState *ttnpb.MACState, multicast bool, maxDownLen uint16, queue []*ttnpb.ApplicationDownlink, downs ...*ttnpb.ApplicationDownlink) (unmatchedQueue, unmatchedDowns []*ttnpb.ApplicationDownlink, err error) {
 	queue, unmatchedQueue = partitionDownlinksBySessionKeyIDEquality(session.SessionKeyID, queue...)
 	downs, unmatchedDowns = partitionDownlinksBySessionKeyIDEquality(session.SessionKeyID, downs...)
 	switch {
@@ -142,8 +143,8 @@ func validateApplicationDownlinks(session ttnpb.Session, macState *ttnpb.MACStat
 
 	for _, down := range downs {
 		switch {
-		case len(down.FRMPayload) > 250:
-			return unmatchedQueue, unmatchedDowns, errApplicationDownlinkTooLong
+		case len(down.FRMPayload) > int(maxDownLen):
+			return unmatchedQueue, unmatchedDowns, errApplicationDownlinkTooLong.WithAttributes("length", len(down.FRMPayload), "max", maxDownLen)
 
 		case down.FCnt < minFCnt:
 			return unmatchedQueue, unmatchedDowns, errFCntTooLow
@@ -172,22 +173,38 @@ func validateApplicationDownlinks(session ttnpb.Session, macState *ttnpb.MACStat
 // - An item's session is neither the device's active session, nor device's pending session;
 // - An item's session matches device's session, but corresponding MACState is missing;
 // - The LoRaWAN version is 1.0.x and an item's FCnt is not higher than the session's NFCntDown.
-func validateQueuedApplicationDownlinks(ctx context.Context, dev *ttnpb.EndDevice, downs ...*ttnpb.ApplicationDownlink) error {
+func validateQueuedApplicationDownlinks(ctx context.Context, dev *ttnpb.EndDevice, fps *frequencyplans.Store, downs ...*ttnpb.ApplicationDownlink) error {
 	if len(downs) == 0 {
 		return nil
 	}
 
-	var err error
+	fp, phy, err := getDeviceBandVersion(dev, fps)
+	if err != nil {
+		return err
+	}
+	var maxDownLen uint16 = 0
+	for _, dr := range phy.DataRates {
+		if n := dr.MaxMACPayloadSize(fp.DwellTime.GetDownlinks()); n > maxDownLen {
+			maxDownLen = n
+		}
+	}
+	if maxDownLen < 8 {
+		log.FromContext(ctx).Error("Data rate MAC payload size limits too low for application downlink to be scheduled")
+		maxDownLen = 0
+	} else {
+		maxDownLen -= 8
+	}
+
 	unmatchedDowns := downs
 	unmatchedQueue := dev.QueuedApplicationDownlinks
 	if dev.Session != nil {
-		unmatchedDowns, unmatchedQueue, err = validateApplicationDownlinks(*dev.Session, dev.MACState, dev.Multicast, unmatchedQueue, unmatchedDowns...)
+		unmatchedDowns, unmatchedQueue, err = validateApplicationDownlinks(*dev.Session, dev.MACState, dev.Multicast, maxDownLen, unmatchedQueue, unmatchedDowns...)
 		if err != nil {
 			return err
 		}
 	}
 	if dev.PendingSession != nil {
-		unmatchedDowns, unmatchedQueue, err = validateApplicationDownlinks(*dev.PendingSession, dev.PendingMACState, dev.Multicast, unmatchedQueue, unmatchedDowns...)
+		unmatchedDowns, unmatchedQueue, err = validateApplicationDownlinks(*dev.PendingSession, dev.PendingMACState, dev.Multicast, maxDownLen, unmatchedQueue, unmatchedDowns...)
 		if err != nil {
 			return err
 		}
@@ -230,7 +247,7 @@ func (ns *NetworkServer) DownlinkQueueReplace(ctx context.Context, req *ttnpb.Do
 				return nil, nil, errDeviceNotFound
 			}
 			dev.QueuedApplicationDownlinks = nil
-			if err := validateQueuedApplicationDownlinks(ctx, dev, req.Downlinks...); err != nil {
+			if err := validateQueuedApplicationDownlinks(ctx, dev, ns.FrequencyPlans, req.Downlinks...); err != nil {
 				return nil, nil, err
 			}
 			dev.QueuedApplicationDownlinks = req.Downlinks
@@ -285,7 +302,7 @@ func (ns *NetworkServer) DownlinkQueuePush(ctx context.Context, req *ttnpb.Downl
 			if dev == nil {
 				return nil, nil, errDeviceNotFound
 			}
-			if err := validateQueuedApplicationDownlinks(ctx, dev, req.Downlinks...); err != nil {
+			if err := validateQueuedApplicationDownlinks(ctx, dev, ns.FrequencyPlans, req.Downlinks...); err != nil {
 				return nil, nil, err
 			}
 			dev.QueuedApplicationDownlinks = append(dev.QueuedApplicationDownlinks, req.Downlinks...)